MUI + Emotion SSR fix and restructuring
