+++ app/(pages)/components/DashboardLayout.tsx
@@ -44,10 +44,17 @@ export default function DashboardLayout({ children }: { children: React.ReactNod
                 <IconButton
                   color="warning"
                   onClick={() => {
-                    // Clear cookie & localStorage then redirect
+                    // Clear cookie
                     document.cookie =
                       "token=; path=/; secure; expires=Thu, 01 Jan 1970 00:00:00 GMT";
-                    localStorage.removeItem("userEmail");
+
+                    // Clear localStorage
+                    localStorage.removeItem("token");
+                    localStorage.removeItem("role");
+                    localStorage.removeItem("userId");
+                    localStorage.removeItem("email");
+
+                    // Redirect
                     window.location.href = "/login";
                   }}
                 >
