+++ app/(pages)/components/JobsTable.tsx
