+++ app/(pages)/components/theme.ts
@@ -1,0 +1,14 @@
+import { createTheme } from "@mui/material/styles";
+
+const theme = createTheme({
+  palette: {
+    mode: "light",
+    primary: { main: "#00897b" },   // teal
+    secondary: { main: "#ffb300" }, // amber
+  },
+  typography: {
+    fontFamily: "var(--font-geist-sans), sans-serif",
+  },
+});
+
+export default theme;
