+++ app/(pages)/layout.tsx
@@ -1,7 +1,7 @@
 import type { Metadata } from "next";
 import { Geist, Geist_Mono } from "next/font/google";
-import "./globals.css";
-import 'bootstrap/dist/css/bootstrap.min.css';
+import "../globals.css";
+import "bootstrap/dist/css/bootstrap.min.css";
 import ClientThemeProvider from "./components/ClientThemeProvider";
 
 const geistSans = Geist({
@@ -21,15 +21,11 @@ export const metadata: Metadata = {
 
 export default function RootLayout({
   children,
-}: Readonly<{
-  children: React.ReactNode;
-}>) {
+}: Readonly<{ children: React.ReactNode }>) {
   return (
     <html lang="en">
       <body className={`${geistSans.variable} ${geistMono.variable} antialiased`}>
-        <ClientThemeProvider>
-          {children}
-        </ClientThemeProvider>
+        <ClientThemeProvider>{children}</ClientThemeProvider>
       </body>
     </html>
   );
