+++ app/(pages)/user/scripts/page.tsx
